+++ src/Intersection.java
@@ -4,7 +4,7 @@ public class Intersection extends Point {
 
     // attributs
     private Primitive objet; // référence à l'objet intersecté
-    private float t; // "distance" paramétrique depuis l'origine du rayon
+    private double t; // "distance" paramétrique depuis l'origine du rayon
 
     /**
      * crée une intersection vide
@@ -26,7 +26,7 @@ public class Intersection extends Point {
      * @param t     la distance paramétrique de cet objet par rapport à l'origine du
      *              rayon
      */
-    public Intersection(float x, float y, float z, Primitive objet, float t) {
+    public Intersection(double x, double y, double z, Primitive objet, double t) {
         super(x, y, z);
         this.objet = objet;
         this.t = t;
@@ -66,7 +66,7 @@ public class Intersection extends Point {
      * 
      * @return la distance paramétrique
      */
-    public float getDistance() {
+    public double getDistance() {
         return t;
     }
 
@@ -87,7 +87,7 @@ public class Intersection extends Point {
     public Intensite eclairer(Scene s, Point obs, int niveau, int px, int py) {
         Point inter = new Point(x, y, z);
         Intensite i = new Intensite(0.0f, 0.0f, 0.0f);
-        float r, v, b;
+        double r, v, b;
 
         // lancer les rayons d'ombrage
         Iterator it = s.sourcesIterator();
